+++ common/src/main/java/whocraft/tardis_refined/command/TardisRefinedCommand.java
@@ -4,10 +4,7 @@ import com.mojang.brigadier.CommandDispatcher;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.commands.Commands;
 import whocraft.tardis_refined.TardisRefined;
-import whocraft.tardis_refined.command.sub.CreateCommand;
-import whocraft.tardis_refined.command.sub.InteriorCommand;
-import whocraft.tardis_refined.command.sub.LevelCommand;
-import whocraft.tardis_refined.command.sub.UpgradesCommand;
+import whocraft.tardis_refined.command.sub.*;
 import whocraft.tardis_refined.command.sub.export.ExportDesktopCommand;
 import whocraft.tardis_refined.common.util.Platform;
 
@@ -17,6 +14,7 @@ public class TardisRefinedCommand {
         dispatcher.register(Commands.literal(TardisRefined.MODID).requires(commandSource -> commandSource.hasPermission(Platform.getServer().getOperatorUserPermissionLevel()))
                 .then(InteriorCommand.register(dispatcher))
                 .then(UpgradesCommand.register(dispatcher))
+                .then(SummonCommand.register(dispatcher))
                 .then(Commands.literal("data").then(Commands.literal("export").then(ExportDesktopCommand.register(dispatcher))))
                 .then(LevelCommand.register(dispatcher))
         );