+++ changelog.md
@@ -10,3 +10,6 @@
 - Enhancement: Added tooltips to various screens to indicate which mod added content
 - Enhancement: Players can now sleep in the TARDIS (No spawnpoint setting)
 - Enhancement: Doors can now be placed at 2x2 rendering by shift right-clicking when placing 
+
+#### Commands
+- Added Sub-Command: Summon